+++ package.json
@@ -29,7 +29,6 @@
     "@types/react-dom": "^19.0.0",
     "astro": "^5.5.5",
     "iso-639-1": "^3.1.6",
-    "jspdf": "^4.2.1",
     "locale-emoji": "^0.3.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
