+++ public/locales/de/translation.json
@@ -440,7 +440,6 @@
     "copy_to_clipboard": "In die Zwischenablage kopieren",
     "copied_to_clipboard": "In die Zwischenablage kopiert!",
     "export": "Exportieren",
-    "export_as_pdf": "Als PDF exportieren",
     "export_as_text": "Als Text exportieren",
     "failed_to_copy": "Kopieren fehlgeschlagen",
     "finish_confession": "Beichte beenden",
