+++ public/locales/en/translation.json
@@ -440,7 +440,6 @@
     "copy_to_clipboard": "Copy to Clipboard",
     "copied_to_clipboard": "Copied to clipboard!",
     "export": "Export",
-    "export_as_pdf": "Export as PDF",
     "export_as_text": "Export as Text",
     "failed_to_copy": "Failed to copy",
     "finish_confession": "Finish Confession",
