+++ public/locales/es/translation.json
@@ -393,7 +393,6 @@
     "copy_to_clipboard": "Copiar al portapapeles",
     "copied_to_clipboard": "¡Copiado al portapapeles!",
     "export": "Exportar",
-    "export_as_pdf": "Exportar como PDF",
     "export_as_text": "Exportar como texto",
     "failed_to_copy": "Error al copiar",
     "finish_confession": "Terminar confesión",
