+++ public/locales/it/translation.json
@@ -440,7 +440,6 @@
     "copy_to_clipboard": "Copia negli appunti",
     "copied_to_clipboard": "Copiato negli appunti!",
     "export": "Esporta",
-    "export_as_pdf": "Esporta come PDF",
     "export_as_text": "Esporta come testo",
     "failed_to_copy": "Copia non riuscita",
     "finish_confession": "Concludi la confessione",
