+++ public/locales/pt-BR/translation.json
@@ -440,7 +440,6 @@
     "copy_to_clipboard": "Copiar para a área de transferência",
     "copied_to_clipboard": "Copiado para a área de transferência!",
     "export": "Exportar",
-    "export_as_pdf": "Exportar como PDF",
     "export_as_text": "Exportar como Texto",
     "failed_to_copy": "Falha ao copiar",
     "finish_confession": "Finalizar Confissão",
